requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.2+cpu 
torchvision==0.11.3+cpu
gdown==4.3.0
numpy==1.21.5
pillow==8.4.0
pytelegrambotapi==4.4.0
telebot==0.0.4
urllib3==1.26.8